requirements.txt
# Core framework
langgraph>=0.0.40
langchain>=0.2.0
langchain-community>=0.2.0
langchain-huggingface>=0.0.3

# LLM / embeddings
huggingface-hub>=0.20.0
sentence-transformers>=2.2.0
faiss-cpu>=1.7.4
rank-bm25>=0.2.2

# Frontend
streamlit>=1.30.0

# Data / validation
pydantic>=2.0

# Utilities
python-dotenv>=1.0.0
